% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.anova.R
\name{anova.serp}
\alias{anova.serp}
\title{An ANOVA method for an object of class serp.}
\usage{
\method{anova}{serp}(object, ..., test = c("Chisq", "none"))
}
\arguments{
\item{object}{A serp object.}

\item{...}{additional arguments.}

\item{test}{type of test to be conducted.}
}
\value{
The ANOVA table of a fitted model.
}
\description{
Provides an ANOVA table for comparing two or more serp objects.
}
\examples{
\dontrun{
anova(object,..., test)
}
}
\seealso{
\code{\link{serp}}
}
