% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{coef.serp}
\alias{coef.serp}
\title{Coefficients for a serp object}
\usage{
\method{coef}{serp}(object, ...)
}
\arguments{
\item{object}{An object of class serp.}

\item{...}{additional arguments.}
}
\value{
A vector of model coefficients.
}
\description{
Returns the coefficients of a fitted object of class 'serp'
}
\examples{
# See serp() documentation for examples.
}
\seealso{
\code{\link{serp}}
}
