% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{logLik.serp}
\alias{logLik.serp}
\title{Log-likelihood for a serp object.}
\usage{
\method{logLik}{serp}(object, ...)
}
\arguments{
\item{object}{A "serp" object.}

\item{...}{additional arguments.}
}
\value{
A single value of model log-likelihood
}
\description{
Log-likelihood for a serp object.
}
\examples{
# See serp() documentation for examples.
}
\seealso{
\code{\link{serp}}
}
