% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{predict.serp}
\alias{predict.serp}
\title{Predict method for a serp object.}
\usage{
\method{predict}{serp}(object, type = c("link", "response", "class"), newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class serp.}

\item{type}{any of response, link or terms.}

\item{newdata}{fresh dataset with all relevant variables.}

\item{...}{additional arguments.}
}
\value{
The object returned depends on \code{type}.
}
\description{
Returns the predicted probabilities, link and class
for an object of class "serp".
}
\examples{
# See serp() documentation for examples.
}
\seealso{
\code{\link{serp}}
}
