% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{print.errorMetrics}
\alias{print.errorMetrics}
\title{Print method for an object of class errorMetrics}
\usage{
\method{print}{errorMetrics}(x, ...)
}
\arguments{
\item{x}{An object of class \code{errorMetrics}.}

\item{...}{additional arguments.}
}
\value{
No return value
}
\description{
Prints out a vector of calculated error value of fitted model.
}
\seealso{
\code{\link{errorMetrics}}
}
