% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{print.serp}
\alias{print.serp}
\title{Print method for a serp object.}
\usage{
\method{print}{serp}(x, ...)
}
\arguments{
\item{x}{An object of class serp.}

\item{...}{additional arguments.}
}
\description{
Prints the data frame returned by the \code{summary.serp} method.
}
\examples{
# See serp() documentation for examples.

}
\seealso{
\code{\link{serp}}
}
