% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{print.summary.serp}
\alias{print.summary.serp}
\title{Print method for an object of class summary.serp}
\usage{
\method{print}{summary.serp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.serp}.}

\item{...}{additional arguments.}
}
\value{
No return value
}
\description{
Prints out the information supplied via \code{summary.serp} method.
}
\seealso{
\code{\link{serp}}, \code{\link{print.serp}}
}
