% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{summary.serp}
\alias{summary.serp}
\title{Summary method for a serp object.}
\usage{
\method{summary}{serp}(object, ...)
}
\arguments{
\item{object}{An object of class serp.}

\item{...}{Not used. Additional summary arguments.}
}
\value{
an object of class "summary.serp", a list (depending on the type of
\code{slope} used). All components from object are already defined
in the \code{\link{serp}} function.
\describe{
  \item{coefficients}{the matrix of coefficients, standard errors,
        z-values and p-values.}
  \item{penalty}{list of penalization information obtained with
        \code{slope} set to "penalize".}
  \item{expcoefs}{the exponentiated coefficients.}
}
}
\description{
Summarizes the results of the fitted model in a dataframe.
}
\examples{
# See serp() documentation for examples.
}
\seealso{
\code{\link{serp}}
}
