% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{AIC.serp}
\alias{AIC.serp}
\title{AIC for a serp object.}
\usage{
\method{AIC}{serp}(object, ..., k = 2)
}
\arguments{
\item{object}{An object of class serp.}

\item{...}{additional arguments.}

\item{k}{fixed value equal to 2.}
}
\value{
A single value of model AIC.
}
\description{
AIC for a serp object.
}
\examples{
# See serp() documentation for examples.
}
\seealso{
\code{\link{serp}}
}
