% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{BIC.serp}
\alias{BIC.serp}
\title{BIC for a serp object.}
\usage{
\method{BIC}{serp}(object, ...)
}
\arguments{
\item{object}{An object of class serp.}

\item{...}{additional arguments.}
}
\value{
A single value of model BIC.
}
\description{
BIC for a serp object.
}
\examples{
# See serp() documentation for examples.
}
\seealso{
\code{\link{serp}}
}
