% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{BIC.serp}
\alias{BIC.serp}
\title{BIC for an object of class 'serp'}
\usage{
\method{BIC}{serp}(object, ...)
}
\arguments{
\item{object}{An object of class 'serp'.}

\item{...}{additional arguments.}
}
\value{
A single numeric value of the model.
}
\description{
Returns the bayesian information criterion of a fitted
object of class 'serp'
}
\examples{
# See serp() documentation for examples.
}
\seealso{
\code{\link{serp}}
}
