% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{coef.serp}
\alias{coef.serp}
\title{Coefficients for a serp object.}
\usage{
\method{coef}{serp}(object, ...)
}
\arguments{
\item{object}{A "serp" object.}

\item{...}{additional arguments.}
}
\value{
A vector of model coefficients.
}
\description{
Coefficients for a serp object.
}
\examples{
# See serp() documentation for examples.
}
\seealso{
\code{\link{serp}}
}
