% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.confint.R
\name{confint.serp}
\alias{confint.serp}
\title{Confidence interval for a serp object}
\usage{
\method{confint}{serp}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class serp.}

\item{parm}{unused argument.}

\item{level}{significance level.}

\item{...}{additional arguments.}
}
\value{
The confidence intervals of a fitted model.
}
\description{
Provides the confidence interval of estimates of an object of class serp.
}
\examples{
\dontrun{
confint(object,...)
}
}
\seealso{
\code{\link{serp}}
}
