% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{logLik.serp}
\alias{logLik.serp}
\title{Log-likelihood for a serp object.}
\usage{
\method{logLik}{serp}(object, ...)
}
\arguments{
\item{object}{An object of class serp.}

\item{...}{additional arguments.}
}
\value{
A single numeric value of model log-likelihood
}
\description{
Returns the Log-likelihood for a fitted object of class 'serp'
}
\examples{
# See serp() documentation for examples.
}
\seealso{
\code{\link{serp}}
}
