% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{print.serp}
\alias{print.serp}
\title{Print method for a fitted serp object}
\usage{
\method{print}{serp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{serp}.}

\item{...}{additional arguments.}
}
\value{
No return value
}
\description{
Prints out a vector of coefficients of the fitted model with some
additional goodness-of-fit measures.
}
\seealso{
\code{\link{serp}}, \code{\link{print.summary.serp}}
}
