% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{print.summary.serp}
\alias{print.summary.serp}
\title{Print method for object of class 'summary.serp'}
\usage{
\method{print}{summary.serp}(x, ...)
}
\arguments{
\item{x}{An object of class 'summary.serp'.}

\item{...}{additional arguments.}
}
\value{
No return value
}
\description{
Prints the data frame returned by the \code{summary.serp} method.
}
\examples{
# See serp() documentation for examples.

}
\seealso{
\code{\link{serp}}
}
