% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.control.R
\name{serp.control}
\alias{serp.control}
\title{Control parameters for serp fit.}
\usage{
serp.control(maxits = 5e01, eps = 1e-07, maxpen = 1e5,
                    minP = .Machine$double.eps, nFold = 5e0)
}
\arguments{
\item{maxits}{the maximum number of iteration.}

\item{eps}{threshold value during optimization at which the iteration routine terminates. In other words, when the reported change in the log-likelihood goes below this threshold, convergence is achieved.}

\item{maxpen}{the upper end point of the interval from zero to be searched for a minimum tuning parameter.}

\item{minP}{A near zero minimum value the fitted probabilities are allowed to get during iteration to prevent numerical instability .}

\item{nFold}{the number of k-fold cross validation for the "cv" \code{tuning} method, with default value k = 5.}
}
\description{
Default control parameters for serp fit. User supplied control parameters could be
specified in the main function.
}
\examples{
# See serp() documentation for examples.

}
\seealso{
\code{\link{serp}}
}
