% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.method.R
\name{summary.serp}
\alias{summary.serp}
\title{Summary method for a serp object.}
\usage{
\method{summary}{serp}(object, ...)
}
\arguments{
\item{object}{An object of class serp.}

\item{...}{Not used. Additional summary arguments.}
}
\value{
A data frame of the fitted model summary.
}
\description{
Summarizes the results of the fitted model in a dataframe.
}
\examples{
# See serp() documentation for examples.
}
\seealso{
\code{\link{serp}}
}
