% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.vcov.R
\name{vcov.serp}
\alias{vcov.serp}
\title{Variance covariance matrix for a serp object}
\usage{
\method{vcov}{serp}(object, ...)
}
\arguments{
\item{object}{An object of class serp.}

\item{...}{additional arguments.}
}
\value{
A variance covariance matrix of a fitted model.
}
\description{
Provides the Variance covariance matrix of an object of class serp.
}
\examples{
\dontrun{
vcov(object,...)
}
}
\seealso{
\code{\link{serp}}
}
